hf-deploy/repo/cv_module/requirements.txt
# ── CV Core ───────────────────────────────────────────────
torch==2.1.0                     # versi stabil, fix DLL + numpy issue
torchvision==0.16.0
transformers==4.35.2             # kompatibel dengan torch 2.1.0
numpy==1.26.4                    # <2.0 wajib untuk torch 2.1.0
Pillow>=10.4.0
opencv-python-headless>=4.10.0

# ── CLIP ──────────────────────────────────────────────────
open-clip-torch>=2.26.1
timm==0.9.16                     # versi stabil untuk open-clip

# ── Object Detection ──────────────────────────────────────
ultralytics>=8.2.0

# ── OCR ───────────────────────────────────────────────────
pytesseract>=0.3.13
easyocr>=1.7.1

# ── Image utils ───────────────────────────────────────────
imageio>=2.34.0
scikit-image>=0.24.0

# ── API ───────────────────────────────────────────────────
fastapi==0.112.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0

# ── MLOps ─────────────────────────────────────────────────
mlflow==2.15.1

# ── Utils ─────────────────────────────────────────────────
loguru==0.7.2
python-dotenv==1.0.1
httpx==0.27.0
aiofiles>=24.1.0

# ── Testing ───────────────────────────────────────────────
pytest>=8.3.2
pytest-asyncio>=0.23.8
